+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/CayenneResourceEntityExt.java
@@ -1,28 +1,11 @@
 package io.agrest.cayenne.processor;
 
-import org.apache.cayenne.query.SelectQuery;
+import org.apache.cayenne.query.FluentSelect;
 
 /**
- * Cayenne extensions of a {@link io.agrest.ResourceEntity} stored in entity properties that tags the entity as
- * Cayenne-managed and provides Cayenne-related entity properties, such as query, etc.
- *
- * @since 4.8
+ * @since 5.0
  */
-public class CayenneResourceEntityExt<T> {
+public interface CayenneResourceEntityExt {
 
-    private SelectQuery<T> select;
-
-    // TODO: generics here are not accurate. The actual query is a column query for nested entities and an
-    //  object query for root
-    public SelectQuery<T> getSelect() {
-        return select;
-    }
-
-    public void setSelect(SelectQuery<T> select) {
-        this.select = select;
-    }
-
-    public boolean hasSelect() {
-        return select != null;
-    }
+    FluentSelect<?> getSelect();
 }