+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/ICayenneQueryAssembler.java
@@ -1,29 +1,27 @@
 package io.agrest.cayenne.processor;
 
-import io.agrest.AgObjectId;
 import io.agrest.NestedResourceEntity;
-import io.agrest.ResourceEntity;
 import io.agrest.runtime.processor.select.SelectContext;
 import org.apache.cayenne.exp.property.Property;
-import org.apache.cayenne.query.SelectQuery;
+import org.apache.cayenne.query.ColumnSelect;
+import org.apache.cayenne.query.ObjectSelect;
 
 import java.util.Iterator;
-import java.util.List;
 
 /**
  * @since 3.7
  */
 public interface ICayenneQueryAssembler {
 
-    <T> SelectQuery<T> createRootQuery(SelectContext<T> context);
+    <T> ObjectSelect<T> createRootQuery(SelectContext<T> context);
 
-    <T> SelectQuery<T> createRootIdQuery(ResourceEntity<T> entity, AgObjectId rootId);
+    <T> ColumnSelect<Object[]> createQueryWithParentQualifier(NestedResourceEntity<T> entity);
 
-    <T> SelectQuery<T> createQueryWithParentQualifier(NestedResourceEntity<T> entity);
+    <T, P> ColumnSelect<Object[]> createQueryWithParentIdsQualifier(
+            NestedResourceEntity<T> entity,
+            Iterator<P> parentData);
 
-    <T, P> SelectQuery<T> createQueryWithParentIdsQualifier(NestedResourceEntity<T> entity, Iterator<P> parentData);
-
-    <T> List<Property<?>> queryColumns(NestedResourceEntity<T> entity);
+    <T> Property<?>[] queryColumns(NestedResourceEntity<T> entity);
 
 }