+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/update/CayenneApplyServerParamsStage.java
@@ -5,6 +5,7 @@ import io.agrest.EntityParent;
 import io.agrest.EntityUpdate;
 import io.agrest.NestedResourceEntity;
 import io.agrest.ResourceEntity;
+import io.agrest.RootResourceEntity;
 import io.agrest.cayenne.path.IPathResolver;
 import io.agrest.cayenne.path.PathOps;
 import io.agrest.cayenne.persister.ICayennePersister;
@@ -82,22 +83,38 @@ public class CayenneApplyServerParamsStage implements Processor<UpdateContext<?>
         // TODO: should we only care about response constraints after the commit?
         constraintsHandler.constrainResponse(entity, null);
 
-        tagCayenneEntities(context.getEntity());
+        tagRootEntity(context.getEntity());
     }
 
-    private void tagCayenneEntities(ResourceEntity<?> entity) {
+    private void tagRootEntity(RootResourceEntity<?> entity) {
         if (entityResolver.getObjEntity(entity.getName()) != null) {
-            CayenneProcessor.getOrCreateCayenneEntity(entity);
+            CayenneProcessor.getOrCreateRootEntity(entity);
         }
 
         if (entity.getMapBy() != null) {
             for (NestedResourceEntity<?> child : entity.getMapBy().getChildren().values()) {
-                tagCayenneEntities(child);
+                tagNestedEntity(child);
             }
         }
 
         for (NestedResourceEntity<?> child : entity.getChildren().values()) {
-            tagCayenneEntities(child);
+            tagNestedEntity(child);
+        }
+    }
+
+    private void tagNestedEntity(NestedResourceEntity<?> entity) {
+        if (entityResolver.getObjEntity(entity.getName()) != null) {
+            CayenneProcessor.getOrCreateNestedEntity(entity);
+        }
+
+        if (entity.getMapBy() != null) {
+            for (NestedResourceEntity<?> child : entity.getMapBy().getChildren().values()) {
+                tagNestedEntity(child);
+            }
+        }
+
+        for (NestedResourceEntity<?> child : entity.getChildren().values()) {
+            tagNestedEntity(child);
         }
     }