+++ agrest-cayenne/src/test/java/io/agrest/cayenne/processor/CayenneQueryAssemblerTest.java
@@ -7,9 +7,10 @@ import io.agrest.base.protocol.Sort;
 import io.agrest.cayenne.cayenne.main.E1;
 import io.agrest.cayenne.unit.CayenneNoDbTest;
 import io.agrest.runtime.processor.select.SelectContext;
-import org.apache.cayenne.query.SelectQuery;
+import org.apache.cayenne.query.ObjectSelect;
 import org.junit.jupiter.api.Test;
 
+import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
@@ -22,49 +23,46 @@ public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
         SelectContext<E1> c = new SelectContext<>(E1.class);
         c.setEntity(entity);
 
-        SelectQuery<E1> q1 = queryAssembler.createRootQuery(c);
-        assertEquals(1, q1.getOrderings().size());
-        assertEquals(E1.NAME.asc(), q1.getOrderings().get(0));
+        ObjectSelect<E1> q1 = queryAssembler.createRootQuery(c);
+        assertEquals(asList(E1.NAME.asc()), q1.getOrderings());
 
         entity.getOrderings().add(new Sort("name", Dir.DESC));
-        SelectQuery<E1> q2 = queryAssembler.createRootQuery(c);
-        assertEquals(2, q2.getOrderings().size());
-        assertEquals(E1.NAME.asc(), q2.getOrderings().get(0));
-        assertEquals(E1.NAME.desc(), q2.getOrderings().get(1));
+        ObjectSelect<E1> q2 = queryAssembler.createRootQuery(c);
+        assertEquals(asList(E1.NAME.asc(), E1.NAME.desc()), q2.getOrderings());
     }
 
     @Test
     public void testCreateRootQuery_Pagination() {
 
-        RootResourceEntity<E1> resourceEntity = new RootResourceEntity<>(getAgEntity(E1.class));
-        resourceEntity.setFetchLimit(10);
-        resourceEntity.setFetchOffset(0);
+        RootResourceEntity<E1> entity = new RootResourceEntity<>(getAgEntity(E1.class));
+        entity.setFetchLimit(10);
+        entity.setFetchOffset(0);
 
         SelectContext<E1> c = new SelectContext<>(E1.class);
-        c.setEntity(resourceEntity);
+        c.setEntity(entity);
 
-        SelectQuery<E1> q1 = queryAssembler.createRootQuery(c);
+        ObjectSelect<E1> q1 = queryAssembler.createRootQuery(c);
 
         assertEquals(10, q1.getPageSize(), "Pagination in the query for paginated request is expected");
-        assertEquals(0, q1.getFetchOffset());
-        assertEquals(0, q1.getFetchLimit());
+        assertEquals(0, q1.getOffset());
+        assertEquals(0, q1.getLimit());
 
-        resourceEntity.setFetchLimit(0);
-        resourceEntity.setFetchOffset(0);
-        CayenneProcessor.getOrCreateCayenneEntity(resourceEntity).setSelect(null);
+        entity.setFetchLimit(0);
+        entity.setFetchOffset(0);
+        CayenneProcessor.getOrCreateRootEntity(entity).setSelect(null);
 
-        SelectQuery<E1> q2 = queryAssembler.createRootQuery(c);
+        ObjectSelect<E1> q2 = queryAssembler.createRootQuery(c);
         assertEquals(0, q2.getPageSize());
-        assertEquals(0, q2.getFetchOffset());
-        assertEquals(0, q2.getFetchLimit());
+        assertEquals(0, q2.getOffset());
+        assertEquals(0, q2.getLimit());
 
-        resourceEntity.setFetchLimit(0);
-        resourceEntity.setFetchOffset(5);
+        entity.setFetchLimit(0);
+        entity.setFetchOffset(5);
 
-        SelectQuery<E1> q3 = queryAssembler.createRootQuery(c);
+        ObjectSelect<E1> q3 = queryAssembler.createRootQuery(c);
         assertEquals(0, q3.getPageSize());
-        assertEquals(0, q3.getFetchOffset());
-        assertEquals(0, q3.getFetchLimit());
+        assertEquals(0, q3.getOffset());
+        assertEquals(0, q3.getLimit());
     }
 
     @Test
@@ -75,12 +73,12 @@ public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
         c.setEntity(entity);
 
         entity.andQualifier(Exp.simple("name = 'X'"));
-        SelectQuery<E1> q1 = queryAssembler.createRootQuery(c);
-        assertEquals(E1.NAME.eq("X"), q1.getQualifier());
+        ObjectSelect<E1> q1 = queryAssembler.createRootQuery(c);
+        assertEquals(E1.NAME.eq("X"), q1.getWhere());
 
         entity.andQualifier(Exp.simple("name in ('a', 'b')"));
-        SelectQuery<E1> q2 = queryAssembler.createRootQuery(c);
-        assertEquals(E1.NAME.eq("X").andExp(E1.NAME.in("a", "b")), q2.getQualifier());
+        ObjectSelect<E1> q2 = queryAssembler.createRootQuery(c);
+        assertEquals(E1.NAME.eq("X").andExp(E1.NAME.in("a", "b")), q2.getWhere());
     }
 
     @Test
@@ -90,24 +88,24 @@ public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
         c.setId(1);
         c.setEntity(getResourceEntity(E1.class));
 
-        SelectQuery<E1> s1 = queryAssembler.createRootQuery(c);
+        ObjectSelect<E1> s1 = queryAssembler.createRootQuery(c);
         assertNotNull(s1);
-        assertSame(E1.class, s1.getRoot());
+        assertSame(E1.class, s1.getEntityType());
     }
 
     @Test
     public void testCreateRootQuery_ById_WithQuery() {
-        SelectQuery<E1> select = new SelectQuery<>(E1.class);
+        ObjectSelect<E1> select = ObjectSelect.query(E1.class);
 
         SelectContext<E1> c = new SelectContext<>(E1.class);
         c.setId(1);
         c.setEntity(getResourceEntity(E1.class));
 
-        CayenneProcessor.getOrCreateCayenneEntity(c.getEntity()).setSelect(select);
+        CayenneProcessor.getOrCreateRootEntity(c.getEntity()).setSelect(select);
 
-        SelectQuery<E1> s2 = queryAssembler.createRootQuery(c);
+        ObjectSelect<E1> s2 = queryAssembler.createRootQuery(c);
         assertNotNull(s2);
         assertNotSame(select, s2);
-        assertSame(E1.class, s2.getRoot());
+        assertSame(E1.class, s2.getEntityType());
     }
 }