+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@
 * #521 Cayenne entity ID part represented by a DB column to be prefixed with "db:"
 * #524 Align APIs for passing ID to requests: "byId" vs "id"
 * #525 UnrelateBuilder
+* #526 Replace deprecated Cayenne SelectQuery with ObjectSelect/ColumnSelect
 
 ## Release 4.9